contrib/serverless/requirements.txt
PyYAML
python-decouple
boto3
git+https://github.com/multi-cloud-explorer/mce-lib-aws.git